backend/requirements.txt
django-allauth[mfa,socialaccount]>=65.3.1
qrcode >= 7.0.0
psycopg2-binary
djangorestframework
dj-rest-auth
meilisearch